requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.5
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.5.2
Flask-WTF==1.1.1
python-dotenv==1.0.0
requests==2.31.0
stripe==7.0.0